Fix set-but-not-used warnings from latest clang

+++ tests/tools/selectv.c
@@ -63,7 +63,7 @@ unsigned int myrand(void)
 
     if (first) {
         seed = time(NULL);
-        adds = (unsigned int)pthread_self();
+        adds = (unsigned int)(intptr_t) pthread_self();
         first = 0;
     }
